requirements.txt
beautifulsoup4==4.14.3
Django==4.2.27
django-apscheduler==0.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
finance-datareader==0.9.100
lxml==6.0.2
pandas==2.3.3
pyupbit==0.2.34
requests==2.32.5
requests-file==3.0.1
websockets==15.0.1
anyio==4.12.0
httpcore==1.0.9
httpx==0.28.1
python-telegram-bot==22.5
